Move entire main.rs into cli.rs
This is the smallest change to make things still compile. Refactoring
comes in the next commit.
