+++ mountpoint-s3/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod autoconfigure;
+mod build_info;
 mod checksums;
+pub mod cli;
 pub mod data_cache;
 pub mod fs;
 pub mod fuse;
